+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
-## [Unreleased](https://github.com/dalance/cargo-trend/compare/v0.2.5...Unreleased) - ReleaseDate
+## [Unreleased](https://github.com/dalance/cargo-trend/compare/v0.3.0...Unreleased) - ReleaseDate
+
+## [v0.3.0](https://github.com/dalance/cargo-trend/compare/v0.2.5...v0.3.0) - 2020-11-25
 
 * [Fixed] wrong count of transitive
 * [Fixed] feature trace at transitive count